+++ js/model3D/shared/Canvas3D.svelte
@@ -22,7 +22,7 @@
 	onMount(() => {
 		const initViewer = async (): Promise<void> => {
 			BABYLON_VIEWER = await import("@babylonjs/viewer");
-			BABYLON_VIEWER.createViewerForCanvas(canvas, {
+			BABYLON_VIEWER.CreateViewerForCanvas(canvas, {
 				clearColor: clear_color,
 				useRightHandedSystem: true,
 				animationAutoPlay: true,