+++ js/checkboxgroup/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/checkboxgroup
 
+## 0.7.1-dev.0
+
+### Dependency updates
+
+- @gradio/atoms@0.18.2-dev.0
+- @gradio/utils@0.10.3-dev.0
+- @gradio/statustracker@0.12.0-dev.0
+
 ## 0.7.0
 
 ### Dependency updates