+++ js/sketchbox/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @gradio/sketchbox
 
+## 0.6.18-dev.0
+
+### Dependency updates
+
+- @gradio/atoms@0.18.2-dev.0
+- @gradio/utils@0.10.3-dev.0
+- @gradio/column@0.2.2
+- @gradio/statustracker@0.12.0-dev.0
+
 ## 0.6.17
 
 ### Dependency updates